requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.31.0
chromadb==0.4.18
sentence-transformers>=2.7.0
huggingface_hub>=0.20.0
requests==2.31.0
pydantic==2.5.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 